data_pipeline(airflow)/jj/requirements.txt
# Additional Python packages for Airflow DAGs
# 
# 설치 방법:
# .env 파일의 _PIP_ADDITIONAL_REQUIREMENTS에 이 파일 경로를 지정:
# _PIP_ADDITIONAL_REQUIREMENTS=/opt/airflow/requirements.txt
#
# 또는 직접 패키지 목록 지정:
# _PIP_ADDITIONAL_REQUIREMENTS=pandas>=2.0.0 numpy>=1.24.0 matplotlib>=3.7.0 torch>=2.0.0 scikit-learn>=1.3.0 optuna>=3.3.0 psycopg2-binary>=2.9.0
#
# 현재 설치된 버전 (2025-11-17 기준):
# pandas==2.2.3
# numpy==1.26.4
# matplotlib==3.10.7
# torch==2.9.1
# scikit-learn==1.7.2
# optuna==4.6.0
# psycopg2-binary==2.9.10

# 데이터 처리 및 분석
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0

# 머신러닝 및 딥러닝
# torch는 CPU 버전으로 설치 (스크립트에서 별도로 설치)
# torch>=2.0.0,<3.0.0  # CPU 버전으로 별도 설치
tensorflow>=2.15.0,<2.17.0  # TensorFlow 2.15.x (Python 3.12, NumPy 1.26.x 호환)
scikit-learn>=1.3.0,<2.0.0

# 하이퍼파라미터 최적화
optuna>=3.3.0,<5.0.0

# 시각화 (선택 사항)
matplotlib>=3.7.0,<4.0.0

# 데이터베이스 연결
psycopg2-binary>=2.9.0,<3.0.0  # PostgreSQL (PostgresHelper에서 사용)
oracledb>=1.0.0,<3.0.0  # Oracle (OracleHelper에서 사용)

# Airflow Providers
apache-airflow-providers-google  # Google Cloud Services
apache-airflow-providers-amazon  # AWS Services
apache-airflow-providers-microsoft-azure  # Azure Services
apache-airflow-providers-influxdb  # InfluxDB
apache-airflow-providers-mongo  # MongoDB
apache-airflow-providers-postgres
apache-airflow-providers-microsoft-mssql>=4.0.0,<5.0.0  # MSSQL (MSSQLHelper에서 사용)
apache-airflow-providers-oracle  # Oracle (OracleHelper에서 사용)
apache-airflow-providers-mysql  # MySQL (MySQLHelper에서 사용)
apache-airflow-providers-common-sql  # Common SQL utilities

# 추가 유틸리티
influxdb-client  # InfluxDB 클라이언트
pendulum  # 날짜/시간 처리 라이브러리

# dbt 및 Astronomer Cosmos 통합 (무료 오픈소스)
astronomer-cosmos>=1.0.0,<2.0.0  # Airflow와 dbt 통합 라이브러리
dbt-core>=1.7.0,<1.8.0  # dbt Core 1.7.x (오픈소스)
dbt-postgres>=1.7.0,<1.8.0  # PostgreSQL 어댑터 1.7.x

# urllib3 호환성 문제 해결 (pyopenssl과의 호환성)
urllib3==1.26.18  # urllib3 1.26.20+는 PROTOCOL_TLS 제거로 pyopenssl과 호환성 문제 발생
# 다른 데이터베이스 사용 시 아래 중 선택:
# dbt-snowflake>=1.5.0,<2.0.0  # Snowflake 사용 시
# dbt-bigquery>=1.5.0,<2.0.0   # BigQuery 사용 시
# dbt-redshift>=1.5.0,<2.0.0   # Redshift 사용 시
# dbt-mysql>=1.5.0,<2.0.0      # MySQL 사용 시

# Note: MySQL 연결은 apache-airflow-providers-mysql이 자동으로 설치함 (mysqlclient, mysql-connector-python 포함)
# aiomysql은 현재 사용하지 않음 (MySqlHook은 동기 방식 사용)

